Update _config.yml
We'll work in a branch, my-pages, that I created for you to get this site looking great. ❇️

Jekyll uses a file titled _config.yml to store settings for your site, your theme, and reusable content like your site title and GitHub handle. You can check out the _config.yml file on the Code tab of your repository.

We need to use a blog-ready theme. For this activity, we will use a theme named "minima".

⌨️ Activity: Configure your site
Browse to the _config.yml file in the my-pages branch.
In the upper right corner, open the file editor.
Add a theme: set to minima so it shows in the _config.yml file as below:
theme: minima
(optional) You can modify the other configuration variables such as title:, author:, and description: to further customize your site.
Commit your changes.